requirements_full_windows.txt
PySide6>=6.7
numpy>=1.26
Pillow>=10.0
opencv-python>=4.9
reportlab>=4.0
pyinstaller>=6.0
pytest>=8.0
ruff>=0.5
black>=24.0
